Cleanup old data-central Person model.
Consolidate Person and User together into one and remove all the stuff
around Contact Owner and List. I think "Contact" was meant to represent
a way for a person to be contacted, e.g. Phone, Email, Address but the naming
made it very confusing. I'm leaving Phone, Email, and Address in place for now
in case we want to use them, but we may end up deleting them too)

Since we're going to be designing the rostering data model from the ground up I
just want to start from something basic. A User, a Role, and the Program they're in
and we'll work up from there.

Consider renaming Program to Course as we start to flesh out the roster data model.

TESTING:
- all specs pass except the Selenium smoke tests for Content Editor.
- able to login and see Content Editor still
- able to load /users route and see a list of users

NOTE:
- the UI around all the models created for the data-central
  project is broken with this commit:
    https://github.com/bebraven/platform/commit/eeb022de85da74928d577a9fa2b8f199a99fadd0
  However, this stage of the project is focused on the models so
  I'm not putting any effort into fixing the UI.