Make the drop table migrations irreversible.
We have a number of models in place in this repo that are a work
in progress and haven't been used in production yet. As we iterate on
those yet to be used models, make it so we can't do a db:rollback of
the corresponding migrations so that when we're finally done and
the models start to be used for real data in production it's clear
from the migrations when that happened. In other words, once we
have real data in production, if we're going to do something that
causes data loss we need to explicitly plan for how to rollback and
restore the data if necessary.

TESTING:
- Tried to rollback 3 steps and it threw IrreversibleMigration
- Dropped the database and recreated it by running the migrations
  from step one to make sure I didn't break a fresh install.