First pass at some grading related data models
Add info about gaurd and live reloading bug to README.

Add bare bones Project model.

A Project has nothing but a name and the ability to view a project.

Add bare bones Lesson model.

Bare bones == a Lesson has nothing but a name and the ability to view it.

Add bare bones LessonSubmission model.

Bare bones == a LessonSubmission has nothing but an association to the user and lesson
it's for and the availability to view it.

Cleanup unused to_show method on bare bones project and lesson model.

Add bare bones ProjectSubmission model.

Bare bones == a ProjectSubmission has nothing but an association to the user and project
it's for and the availability to view it.

Add bare bones CourseModule model.

Bare bones == a CourseModule has nothing but a name and associations program,
project, and lesson.

Nest project and lesson submissions under their parent route.

Add navigation from Programs down to CourseModules.

Add links to the stuff we're working on to the home page.

Showcase routes between program, modules, projects, lessons, and submissions.

Add bare bones points and grades columns on projects and lessons.

Add example code for calculating the final grade for a user.

Add example of displaying the total grade in the UI.

Add some comments about lesson, project, and course_module modals.

Make course_module model skinnier.

We still need to make a decision about where to store services
and ActiveModals that we use to make our models skinnier by abstracting
out the business logic, but for now we have that sort of thing in lib
so that's where I'm putting it.

Better name for method to view grades as a percent

Use parentheses for method definitions.