Rubric related data models.
See this doc for an overview:
https://docs.google.com/document/d/160uyajp_be6V8v2aaoIa6NxrlT4DmXXqa5SJRLaJ3Rc#heading=h.d5ihwqwhakzx

Add bare bones rubric data models

Add associations between projects and bare bones rubric models.

Database migrations to follow.

Database migrations for rubrics data models.

Also, fixup a few of the associations.

Remove rubric name.

We can add one later if we need to, but for this first pass
a project has_one rubric, so the rubric name is the name of the project.

Move rubric models out of subdirectory.

To get this to work, I'd have to either namespace them or
tinker with autoload paths. It's too early to know how our
models are going to shape up so lets deal with organizing them
once it becomes annoying to have them all in one place.