Use a Procfile so we control heroku release behavior.
Rails defaults to Puma now, so we're good with the web command
but may as well make it explicity. However, this addes a "release"
process that runs db:migrate during releases so that it happens
automatically and there is 0 downtime during a deploy that depends
on a database migration to be manually run.