DRY up the index controller methods and exempt non-standard controllers.
CasController and PostalCodes don't use the normal pattern for Rails
CRUD controllers. Exempt them.

Note: I'm not happy with the solution for exempting certain controllers.
I would have liked to use respond_to?() and checked if they had implemented
the appropriate methods so that it's all dynamic, but I couldn't figure out
how b/c when the module is included it's done from the ApplicationController
and we need to check if the sub-class including the module responds. There
must be a way to do that?!?