Rubrics and Grade Categories and Specs (#110)
Squashing a bunch of commits related to rubrics, grade categories, and adding specs / database constraints and merging them back into the grading branch. Here are the commits squashed in this merge:

* Rubrics specs

* Somehow dbrefresh script permission got fubar'd.

* Remove CourseModules and replace with GradeCategory.

Eventually, we may bring the concept of CourseModules back,
but for the purposes of grading it didn't make sense to define
the percent of your total grade that some set of lessons or projects
made up on a CourseModule. The concept of a course module is simply to
group a related set of learning content into a topic or theme, mostly impacting
course navigation. For grading, we want to categorize projects
and lessons and set the percentage of the total grade based on that. E.g.
all lessons related to course engagement would be in an Engagement category,
which may span multiple weeks or modules but the entire category is worth 15%
of your total grade.

Note: this breaks the GradeCalculator which I'll fix and add some specs
in a followup commit. Part of that will be to define a second level of
defining what percent something is worth. E.g. without a GradeCagegory
each project or lesson should have a percent within that category and not
implicitly calculate that using the points. Points for one project/lesson
shouldn't have an impact on it's weight towards your final grade. They are
only used for making an intuitive way to grade that particular project/lesson.

* Make dbrefresh use db:migrate instead of loading from the schema.rb

See comment in code for why.

* Add percent_of_grade_category to Lesson and Project.

For all lesson and projects in a category, the percent of the overall
grade within that category is defined for each.

This adds some super basic specs too.

* Make GradeCalculator raise an exception.

Come back and implement this when we need it. Right now I'm just
focused on the data models and not the actual logic that uses them.

* Add a basic GradeCategory spec.

* Specs for lessons and projects being in a GradeCategory

For now, enforce all lessons and projects being categorized into a
GradeCategory for purposes of grading and determining their weight towards
the final grade.

* Add contraints and specs for LessonSubmission and ProjectSubmission.

* Get rid of organization_spec WARNING.

* Cleanup GradeCategory based on code review feedback.

- Makes projects and lessons belong to Course through GradeCategory
- Adds validation for percent_of_grade related columns to be between 0 and 1