Prototype of user story to allow users to create an account and login.
Exposes URLs like: http://platformweb/users/sign_up?email=blah@blah.org
Presents a signup form to enter a password.
Emails a confirmation link.
After confirmation, redirects to sign in through SSO.
Brings you to platform homepage.

The main thing this commit is aimed at is getting Devise
cas_authenticable and database_authenticable to play nice.
They are only meant to be one or the other. This sort of hacks
around that by keeping the User model cas_authenticable, aka
you login through CAS SSO, but it adds the database_authenticable
module in and creates some controllers, routes and views to allow registration
and confirmation to happen through the database_authenticable
framework.

Tons of TODOs to clean up and the actual flow needs to be ironed out
as well as specs written and fixed.

Test Plan:
- Set the APPLICATION_HOST env var. Then run through the above steps,
  grabbing the confirmation link out of the server logs since the mailer isn't actually setup.

TODOs off the top of my head:
- fix existing specs
- write new specs
- grep for TODO in this commit and do those
- on sign in, look up if they are in the Portal and redirect there if so
- registration should not create admin users
- implement forgot password and trackable
- setup a dev env mailer solution (and verify that a real one works in prod)