Switch account creation user story from email to Salesforce ID.
This implements the ability to create an account using links such as:
  http://platformweb/users/sign_up?u=<salesforce_id>

The user is prompted to enter a password and then we look up their
Contact record in Salesforce and create a User model with the name
and email from Salesforce. They get a confirmation link to that email
which allows them to start using their new account.

Test Plan
1. Create a Contact record in the Salesforce instance that your environment
   is configured to work with.
2. Grab the Contact Id and hit the above URL with it.
3. Enter a password and submit.
4. Grab the confirmation link out of the logs (the mailer is still not setup).
5. Go to that link and you should be able to login with your password.