Enable running the server in either CAS local auth or remote auth modes.
This is controlled by the BZ_AUTH_SERVER environment variable. If set,
the server runs in authentication mode where it calls into our Join
server to verify the credentials b/c that is where the user account lives.
If NOT set, it runs in a mode where it calls into the local database to
authenticate account that have been created there.

We're in a transition period where we are running two servers in two
different modes serving a different population of users. Once we can
switch any/all servers to use local database authentication / accounts,
delete all the stuff that toggles on the BZ_AUTH_SERVER env var b/c this
is dirty, dirty, dirty.

Test Plan:
- Set the BZ_AUTH_SERVER and run all specs. Also, try logging into the portal
  using an account which lives there and in the Join server but not in the
  local platform database.
- Unset the BZ_AUTH_SERVER variable and run all specs again. Try registering
  and logging in as a local account.