Refactor Sync To LMS and account creation to work with Salesforce Programs.
https://app.asana.com/0/1170770467635599/1173648174910505

We have two separate servers running the Portal, one for normal Braven and one
for Braven Booster. There may be conflicts with the course_id. Also, for Booster
we're going to have a "Holding" Program and then split folks out into Program's
specific to the run of the program they're taking. Also, folks may sign-up last
minute and we need to provision their account in Canvas. Lastly, Booster folks will not
be mapped to a Cohort. They will have a scheduled day/time and we need to add them
into Canvas in corresponding sections.

This commit aims to solve all that.

In addition to making Sync To LMS run off the Salesforce Program, it also changes
the account creation / sign_up to create their Canvas account on the fly at the time they
register their account. This was last minute sign-ups will work without running the sync.

Test Plan:
- bundle exec rspec spec
- Fully confirm and enroll a Participant in a Salesforce Program. Try signing up using their
  Contact Id and notice that they get Canvas access.
- Ditto but run Sync To LMS with the Program and see that they get Canvas access and are provisioned
  in a section for their scheduled day/time.