When resending the confirmation, indicate something happened.
I re-did the flow slighly to bring you to a check your email page
with no re-send confirmation link instead of flashing a "link sent"
message b/c the langauge didn't make sense on the page if we just left
it at a flash. This makes the routes a bit more normal.