Fix production loading the right fonts.
Dev uses assets that are compiled and loaded on the fly. Prod
requires all assets to be precompiled. The prior commit allows you
to try out both. This commit simply fixes that path that we load from
to load the correct font regardless of dev vs prod mode.

Test Plan
- Change RAILS_ENV to production and rebuild. Ensure that the sign-in page
  loads the correct fonts
- Do the same in development mode.