Update the layout for account related views.
Puts them all in the root container and section divs. Also, change the title
of the layout to Braven Login. It could be staff or fellows or other and
we don't know if they are trying to login to the platform or the portal or some
future "app", so just keep it vague.