More code review feedback. Don't load React stuff in application pack.
When we start using React outside of Content Editor, we can add it in
but for now removing it reduces the sign-in page size from 235KB to 106KB
and the load time of the page from ~350to 187ms.

Also, prod builds do need to create tmp/pids, but I moved it to the startup
script and out of the docker image.