Add easier way to troubleshoot issues in gems.
In my case, I need to debug some issues in this gem
https://github.com/bebraven/rubycas-server-core so with this change I can
now just go into the container, navigate to the code in `vendor/bundle`
add logs or binding.pry statements, restart the container and the changes
will be run.

Note the comment in the README about needing to blow away the vendor-bundle
volume if you muck it up too much and want to get back to a pristine state.