Enable single-signout setting in SSO server.
With enable_single_sign_out set to true, it calls into each app that has authenticated
with this ticket and tells them to logout.

The `enable_single_sign_out` setting was only enabled in the test
configuration. It should be enabled for all environments. I think this was an
oversight. Refactor the whole rubycas.yml settings file so that shared settings
are defined in one place and each env only changes settings specific to it.

Test Plan:

platform logout
1. Login to platform.
2. Login to portal in a separate tab.
3. Logout of platform.
4. Try to hit a new page in portal, it should redirect to the sign-in page.

portal logout
1. Repeat step 1 & 2 above.
2. Logout of Portal
3. Try to hit a new page in platform, redirects to sign-in page.