On a non Booster server, add a link to the Booster login.
If folks don't bookmark the booster login link and just google
Braven, they'll likely end up at https://bebraven.org and hit
Log In at the top-right. That will bring them to the normal
Accelerator login. On that server, this commit adds a link to login
to the Booster portal if that's what they were looking for.

Test Plan
- Go to the login page with APPLICATION_HOST set to something with
  booster in it and the link shouldn't be there. Change the APPLICATION_HOST
  to something without booster in it and the link to login to Booster should
  show up on the login page.