Implement LTI deep linking and assignment flow (#267)
* Temp commit: LTI assignment placement

* Temp commit: LTI assignment placement

* LtiIdToken now inspects the payload and returns the class to handle it.

This is just a quick and dirty demo to get feedback. Need to clean this
up if we decide to go this route.

Note: the ClassMethods module stuff is me just cleaning that up. Private
class methods don't just work like instance methods.

* Missed a file.

* Add LtiResourceLinkRequest and implement more of DeepLinkingRequest helper.

* Frist pass of implementing LtiDeepLinkingResponseMessage helper.

* Make serialization of id_token_payload transparent.

ActiveRecord will convert to/from JSON so we don't have to.

* Partially fix specs for new target_link_uri logic.

The specs now use the value in the payload and not the parameter
passed. However, it still fails b/c of the state parameter added.
That parameter may be ok for a Deep Link, but the redirect for a
normal launch of a resource should not have it. Need to fix that.

* Moar LtiDeepLinkingResponseMessage tweaks

* Missed some stuff in DeepLinkingResponseMessage

* More DeepLinkResponseMessage cleanup

* Remove old views again

* Add copyright notices to Keypair code

* Use session instead of the state param to look up the current LTI launch info

* Keypairs takes care of rotation and creation. I tested this by updating the created_at col in the db

* Configure CSP

* Update course_contents views

* Remove old iframe code

* Change LtiLaunch to have two states - authenticated vs not.

In the launch flow, use the nonce to match between login and launch in
order to ensure we're grabbing the correct LtiLaunch record. In other places,
when getting the current LtiLaunch that is valid for the session, filter out
anything that may have failed to launch successfully and only get the most
recent successful launch

* Write some controller tests

* Refactor LtiLaunch to use state param instead of browser session.

* Missed some classes in last commit

* Don't hardcode path when selecting an assignment for LTI

* Remvoe unused Keypair.expired method

* Explicitly remove x-frame-options

* Fix the unique state index migration

* Add a trace around the external JWKS request

* Remove empty helper

* Add a helper spec

* Change target uri params

* Remove honeycomb for now since it crashes rspec

Co-authored-by: Brian Sadler <sadler.brian@gmail.com>