Another iteration of LessonContentsController specs.
Undo some of the spec config changes that were breaking stuff.
E.g. we want transactional fixtures turned on so that each test
is isolated and has a clean database. We don't need database cleaner
with this approach. Also, I stubbed out the calls to LessonContentsPublisher
b/c we have separate specs for that and these should focus on the logic
of the controller only.

Test Plan:
bundle exec rspec