Adds a reverse proxy to the AWS S3 hosted files for the LTI lessons.
Now we can give Canvas a URL to launch on our server for the Rise360
lesson content which is hosted on AWS S3 and our server will proxy the
request/response so that both Canvas and the browser think the files
live on our server which we can more easily control the trusting relationship
we've developed.

See: https://app.asana.com/0/1174274412967132/1186566037117865/f

Summary is that we were loading static AWS S3 websites into the browser and the
javascript in those pages was trying to AJAX call into our LRS Proxy in order to
send xApi statements, but since it's cross-origin things went awry. The above
task has more context on the discussion.