Implement Project submissions and the ability to view as a TA (#304)
Note: there are a TON of loose ends in this PR. We've queued up Asana tasks to do a followup PR to fix it up. The goal here is just to get it in a demo-able state and not cause any issues with other features in production.

----
* HAX to get a sense of how we'll show a Project submission to a TA

* Processes the course_contents HTML to show the submission.

Note: you must pass a `student_id` query string parameter in now. E.g.

  https://platformweb/course_contents/3/versions/2?student_id=55

* HAX: working version that pulls from the LRS based on retained-data ID

I'm about to switch it to store all fields in a Project with the same activity ID
and then pull them all and loop through the hash looking at the name of the object
to match it up.

* MOAR HAX: just putting up my broken hacky code to preview for Jess.

Right now, I'm manually hard-coding both a registration ID (aka an id for the course)
and an activity_id (aka an id for the project/assignment). I'm attempting
to use the activity_id to be able to query the LRS and pull all statements
sent for a particular user and a particular project. Then parse that out and populate
the view of the submission for a TA to see (or the student themselves).

The xAPI message is currently breaking and I don't know why. Once I get that to
go through then I need to figure out how to not hard-code this stuff and implement
the code that looks up the most recent thing the user typed per input field
and use that as the defacto value for the field.

* Make xAPI statements for a project all use the assignment URL as the activity ID.

This will allow us to group statements per project when querying them. It's how
Rise360 does lessons, using the same activity ID. We should probably still
configure a "registration" for the statements, but maybe it's not necessary
since the registration is implicitly in the activity as the course #? That's
not true for lessons though, so for consistency we probably want to do it for both.

* Remove TODO from lti_assignment_selection_controller.rb

I think this is how we'll actaully do this. Save a version when we
publish to Canvas.

* HAX to get a sense of how we'll show a Project submission to a TA

* Processes the course_contents HTML to show the submission.

Note: you must pass a `student_id` query string parameter in now. E.g.

  https://platformweb/course_contents/3/versions/2?student_id=55

* HAX: working version that pulls from the LRS based on retained-data ID

I'm about to switch it to store all fields in a Project with the same activity ID
and then pull them all and loop through the hash looking at the name of the object
to match it up.

* MOAR HAX: just putting up my broken hacky code to preview for Jess.

Right now, I'm manually hard-coding both a registration ID (aka an id for the course)
and an activity_id (aka an id for the project/assignment). I'm attempting
to use the activity_id to be able to query the LRS and pull all statements
sent for a particular user and a particular project. Then parse that out and populate
the view of the submission for a TA to see (or the student themselves).

The xAPI message is currently breaking and I don't know why. Once I get that to
go through then I need to figure out how to not hard-code this stuff and implement
the code that looks up the most recent thing the user typed per input field
and use that as the defacto value for the field.

* Make xAPI statements for a project all use the assignment URL as the activity ID.

This will allow us to group statements per project when querying them. It's how
Rise360 does lessons, using the same activity ID. We should probably still
configure a "registration" for the statements, but maybe it's not necessary
since the registration is implicitly in the activity as the course #? That's
not true for lessons though, so for consistency we probably want to do it for both.

* Remove TODO from lti_assignment_selection_controller.rb

I think this is how we'll actaully do this. Save a version when we
publish to Canvas.

* Implement project submission TA view to load values from the LRS.

Fetch all xAPI statements for a project/user in one go, parse them
into a hash that maps the ID of the field (aka retained-data) to the
value.

* Remove TODO. Using state in course_contents_controller#show is fine

* Pin xapi gem to a ref instead of a branch

Prevents Gemfile.lock sync issues if we ever update the gem

* Remove extra custom field

* Parse creds from token

* Update xapi login

* Implement submitting a project & viewing submission on Canvas

 - Fellow can click "Submit" button to submit a project in Canvas
 - TA can use SpeedGrader to view submission in Canvas

More clean up

* Handle no answer + Submit

* Fix bug where it wasn't returning the latest data.

The "ascending" parameter was ignored by the gem's utiity function to
create an LRS query. Initialize the Xapi::StatementsQuery object myself.
Now the last statement processed for a given data_input_id (aka retained data)
will be the most recent one input so the map will set it to that when creating
the hash.

* Fix course_content_histories_controller_spec by stubbing out

* Fix course_contents_controller_spec by stubbing out

* Link Asana task to TODO

* Add more Asana task links to the TODO's

* Remove hard-coded "registration" from Rise350 launch URL.

We have a task to implement this for real. For now just comment
out where we would add the implementation.

* Make old specs pass. We still need to implement specs for this.

* Throw if activity_id of non-assignment launch is used.

Still have to write specs.

* Rename CSS class to highlight user input.

* Cleanup the pre-population of user inputs slightly and add TODO.

* Cleanup setting LRS connection info

* Cleanup lti_score submission slightly.

* Requiring "state" param in course_contents broke preview.

* Get specs to pass.

* Fix submissions. Need proper objects when creating the score.

Co-authored-by: rshipp <github@rshipp.com>
Co-authored-by: Jess Ouyang <jess.ouyang@bebraven.org>