Instrument RestClient calls with Honeycomb tracing and logging. (#321)
* Instrument RestClient calls with Honeycomb tracing and logging.

Task: https://app.asana.com/0/1174274412967132/1187503001085713

If you go under the covers of all RestClient requests, they end up calling
the instance method `execute` on the Request object. This commit prepends
a module that overrides that method to wrap it with a Honeycomb span and
to read the relevant info out of any exceptions that are thrown and trace/log
that as well.

- `bundle exec rspec`
- Do a Project, submit it, make sure data flows to the LRS and is also
  sent to Honeycomb. Here is a sample query: https://ui.honeycomb.io/braven/datasets/stagingplatform.bebraven.org/result/aijLvyA7Fdo
- Change the payload of a message from some RestClient API usage so that
  it will throw an exception. Make sure the exception details come through
  in the logs and on Honeycomb

* Implement specs for Honeycomb instrumentation of RestClient.

Note: using the raw `headers` attr_reader treats `:authorization` and
`Authorization` differently. Using `processed_headers` normalizes them
so that it doesn't matter how the client of the RestClient sets them.

Test Plan:
- `bundle exec rspec spec/lib/restclient_instrumentation_spec.rb`