Fix Project Submission button CSS. (#374)
The project submission button uses Bootstrap CSS. This is made available
in the application webpack and is only currently used for the admin layout.
Add it to the proejct layout. Eventually, we need to clean this up and
include bootstrap everywhere most likely, but for now just add it to
the project layout as well. Note that this works in dev b/c locally there
is the webpacker-dev-server that compiles it:
https://stackoverflow.com/questions/58328026/css-and-bootstrap-not-loading-in-rails-6-app-when-deployed-to-heroku

Test Plan:
- Open an LTI project as a student and scroll to the bottom when the app
  is built and running in production mode. The button should be full width
  and have the `.btn-block` CSS class applied to it.