Initialize LTI assignments linked to Projects on new Course. (#387)
* Initialize LTI assignments linked to Projects on new Course.

When a new Program is launched, we need to adjust the LTI assignments
to launch Project content (and modules) associated with the new course
through BaseCourseCustomContentVersion. This commit refactors the overall
new Course initialize logic into a new service and implements the project
initialization.

TODO: handle modules (aka lessons) as well.

* Fix specs and split fellow and LC course copy in LaunchProgram spec

* Address code review feedback.

Punting on writing specs for the assignment LTI launch URL adjustment
since we're going to refactor. Also, require this to be run off a Canvas
course copied from a CourseTemplate (for now). Eventually if we allow initializing
the local DB from a normal course, for example as a way to create a new template,
we'll need to revisit.

* Cleanup logging