Stop logging LtiLaunch cleanup job in Rails console (#404)
When starting the Rails console, the CleanupLtiLaunchTableJob
would run 30 seconds later and corrupt the terminal prompt right
in the middle of typing commands. This just makes it run immediately in dev.
You don't actually see anything logged b/c this seems to happen before the
logger is all setup and hooked up to STDOUT, but if you drop a "puts"
in the job you'll see that it's still running.

Test Plan:
Connect to the Rails console in dev and see that cursed "Running cleanup
job" doesn't happen 30 seconds later while you're typing messing up the prompt.