Implement update and delete of a Project or Survey (#413)
* Add the UI for creating, updating, and deleting Projects and Surveys.

1. Adds a list of Projects and Surveys to the BaseCourse edit page.
2. Clicking "Add Project" or "Add Survey" shows the list of Projects or Surveys to choose from.
3. Clicking "Create" on the edit page calls the create action on base_course_custom_content_versions (not implemented yet)
4. Clicking on the name of an existing project shows the published version contents in Canvas.
5. Clicking "Publish Latest" on a project calls PUT on the existing BaseCourseContentVersion record (not implemented yet)
6. Clicking "Delete" on a project calls DELETE on the existing BaseCourseContentVersion record (not implemented yet)
7. Clicking on the name of a Course or CourseTemplate from the main page shows a read-only view of
   everything on the Edit page.
8. Only a CourseTemplate shows the button to Edit
9. If you URL hack the edit routes for a course, an exception is thrown.

* Add Asana task for TODO we'll do in a future PR

* More fixing of specs and not breaking the old stuff.

We're going to do a followup PR with the core logic of
create, update, delete for a BaseCourseCustomContentVersion.

Specs are still failing. Decision is to rip out anything failing related
to New or Edit of a Course. But we need to figure out what to do with
New Course Template related failing specs.

* Fix base_courses_controller_spec

We're not longer supporting editing a Course, only a Course Template.
Also, we haven't figured out how we'll initialize a CourseTemplate and
associate it to Canvas. So remove the Edit button for a course and
comment out the New page.

* Fix lti_assignment_selection_controller_spec.rb

* Put back spec I'm not touching in this PR.

The goal of this PR is to get the Course Mgmt UI ready for
implementing create, update, delete for Projects and Surveys.
I'm not aiming to cleanup the convenience views for BaseCourses
that were put in place to make development easier while we worked on
this stuff.

* Implement update and delete of a Project or Survey

* Code review feedback - clearer attribute

* Bad merge and add a TODO link for code review