Enhance DryCrud::Controllers::Nestable to set parent instance variable (#418)
* Enhance DryCrud::Controllers::Nestable to set parent instance variable

The original implementation only took care of filtering the list of models
shown by the index to be those who are associated with the parent_id. This
enhances DryCrud to set the instance variable. E.g. if we have the following
URL:

    https://platformweb/base_course_custom_content_version/1/project_submissions/new

then a @base_course_custom_content_version instance variable will automatically
be set with the parent BaseCourseCustomContentVersion[id=1]

Test Plan:
- bundle exec rspec

* Enhance DryCrud to set parent model instance.

Test Plan
- bundle exec rspec