Remove obsolete TODO. (#443)
When I was working on the refactor here: https://github.com/bebraven/platform/pull/437
the project submission I was trying was failing. There still may be an issue, in which
case we'll uncover it when we do a full end-to-end round of testing from CourseTemplate,
to Program Launch, to enrolling a student, to submitting a project, but when I setup the
local database properly with a student in a section tied to a course with a submission
and tried again, it worked. So I'm removing this TODO for now.