Remove fallback code for backwards compatibility. (#445)
* Remove fallback code for backwards compatibility.

This makes the Sync To LMS user provisioning fail and throw exceptions
if the the local database is missing the course or section that the user
was in or should be in.

We put this fallback code in so that we wouldn't break things in the Portal or
Booster apps since they don't use the locally populated Canvas data. We are no
longer backwards compatible with the old Portal and Booster environments and have
branched. The development of Highlander is evolving separately.

Task: https://app.asana.com/0/1174274412967132/1197893935338145

`bundle exec rspec spec/services/sync_portal_enrollment_for_account_spec.rb`
`bundle exec rspec account_creator_spec.rb`

* Fix spec

Note: these specs seem not so great but it was outside the scope of this
PR to make them better.