Enhance DryCrud to not require setting dry_crud_enabled? false (#452)
* Enhance DryCrud to not require setting dry_crud_enabled? false

The solution here is so simple, just check if the model class is defined
and don't register the before_actions, but i wrestled with trying to get
it to use the `model_class` method from within the `included` context so
I just ended up not being very DRY and just copy/pasting the code that looks
for the model class. Oh the irony.

* Add specs for DryCrud

TODO: specs for DryCrud Nestable

* Make DryCrud work on the subclasses if it was included in a base class

We include DryCrud on the base ApplicationController but it's logic needs to
apply for the subclasses. This change listens for when the class that DryCrud
is included in is inherited from and defines the `before_actions` on the
subclass if necessary.

* Move DryCrud to the Concerns dir

* Only run DryCrud on actual ActiveRecord::Base models.

The CasController ended up have subclasses for various
Devise related controllers, one of which had a valid class
defined to match the controller name (aka Devise) but which wasn't
actually a model.

* Better naming

* Fix check whether model inherits from ActiveRecord::Base

We have the class, not an instance of the class. The `<=` operator
checks if a class inherits from or is the other class.