Add Peer Reviews to Course Management page (#473)
* Hax. Add Peer Reviews to Course Management page

* Rebase + specs

* Cleanup PeerReviews and Publishable a bit.

I wanted to be able to detect when a Publishable controller was missing
the required behavior on the controller to make it work. So I refactored
things from instance variables to methods and added a `method_missing`
handler to raise a nicer error message about the problem if you forget.

Also, some of the comments were taken from the Submittable concern that
didn't apply here. We're not doing meta-programming so you could use this
without a `Nestable` module. You just need a `BaseCourse`.

* Fix the Peer Reviews label wrapping over two lines.

Just need a bit more space.

* Another small UI tweak to put a space between the Course Resources and Add button.

Make the Course ID field line up too

Co-authored-by: Brian Sadler <sadler.brian@gmail.com>